md_07/HW/pyproject.toml
[tool.poetry]
name = "homework 7"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
alembic = "^1.11.3"
sqlalchemy = "^2.0.20"
psycopg2 = "^2.9.7"
tabulate = "^0.9.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
faker = "^19.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
